Added dependencies package
I went ahead and made a (hopefully) simple dependency injection container which can be used to register singletons, lazy singletons and factories with optional names and parameters.

I'm sharing here in hopes it will be useful for the Dart community. This is especially useful for Flutter, where mirrors are disabled.